decomp: reloc-masked wrong-symbol fixes — rb3 (37 functions)
- HasCorrectPlayerCount__19AccomplishmentPanelFv: DIFFERENT_FUNCTION / wrong-callee sibling-overload (100.0->100.0)
- LoadStickerTex__8PatchDirFP12PatchStickerb: outline-wrapper-linkage-mismatch (99.8->99.9)
- getConfig__12ECAsyncOpEnvFv: DIFFERENT_FUNCTION (99.5->99.6)
- SetState__15SaveLoadManagerFQ215SaveLoadManager5State: DIFFERENT_FUNCTION (99.8->99.8)
- CreateVFCache__11CacheMgrWiiFv: DIFFERENT_FUNCTION (99.2->99.2)
- Poll__10WiiContentFv: DIFFERENT_FUNCTION / wrong-callee via MW struct-tag name-man (99.8->99.9)
- DrawCursor__15CharClipDisplayFv: Reloc-masked wrong-symbol (99.5->99.7)
- LoadSubDir__9ObjectDirFiRC8FilePathR9BinStreamb: DIFFERENT_FUNCTION (99.4->99.4)
- BuildProfileUploadOps__14EntityUploaderFRPP10EntityDataP11BandProfile: outline-wrapper-mismatch (99.1->99.2)
- Init__6LocaleFv: DIFFERENT_FUNCTION (98.9->99.1)
- FillSetlistWithAccomplishmentSongs__19AccomplishmentPanelF6Symboli: outlined-function (98.5->98.5)
- PollSearch__11CacheMgrWiiFv: wrong-callee (98.1->99.9)
- Swing__22JoypadTrackWatcherImplFibb11GemHitFlags: wrong-callee (98.1->98.1)
- OpenFile__Q35Movie4Impl11MovieLoaderFv: DIFFERENT_FUNCTION (97.5->97.5)
- Set3DCharAll__10WorldCrowdFv: outline-wrapper removed / direct-call restored (97.6->97.6)
- BuildDrawState__9UIListDirCFR21UIListWidgetDrawStateRC11UIListStateQ211UIComponent5Statef: DIFFERENT_FUNCTION callee correction (97.3->97.3)
- NotifyPlayerOfAccomplishment__22AccomplishmentProgressF6SymbolPCc: DIFFERENT_FUNCTION — compiler-outlined trampoline call targe (97.3->97.3)
- SplitFromBack__4HeapFi: improved (96.9->97.0)
- Poll__5TrackFf: reloc-masked wrong-symbol / call-scheduling (96.6->100.0)
- PrintoutSaveSizeInfo__15SaveLoadManagerFv: DIFFERENT_FUNCTION — source-introduced _outline_Str (96.3->96.4)
- Poll__15SaveLoadManagerFv: wrong-callee (95.6->95.9)
- LoadMainCharacters__12BandWardrobeFP11BandCamShot: wrong-macro (94.6->94.6)
- micSetMute: DIFFERENT_FUNCTION (94.1->94.2)
- DebugDownload__20StoreMetadataManagerFv: wrong-forward-decl-signature (93.3->93.3)
- micOpen: wrong-namespace-scope-for-type (93.5->93.6)
- TryAddFace__Q213BandPatchMesh9WorkVertsFii: DIFFERENT_FUNCTION (89.2->89.2)
- Reskin__13RndMeshDeformFP10SyncMeshCBb: DIFFERENT_FUNCTION / wrong-callee (87.5->88.3)
- __dt__12BudgetScreenFv: real (83.0->99.7)
- Library_OnGetVolDone__6usbmicFlPv: control-flow-shape (77.1->79.1)
- __AXNextFrame: missing-calls (72.3->81.9)
- getParentalControlInfo__12ECAsyncOpEnvFv: DIFFERENT_FUNCTION - wrong forward-declared parameter type f (64.0->64.1)
- StartSearching__17RVSessionSearcherFP4UserRC14SearchSettings: decomp stub (58.8->80.3)
- ExtractCallSpecificResults__Q26Quazal32NintendoManagementProtocolClientFPQ26Quazal7MessagePQ26Quazal19ProtocolCallContext: DIFFERENT_FUNCTION x5 (44.4->73.9)
- compare__Q23std55basic_string<c,Q23std14char_traits<c>,14ECAllocator<c>>CFPCc: wrong-callee (42.8->47.0)
- getSerialNumber__12ECAsyncOpEnvFv: wrong-callee overload mismatch (0.0->37.6)
- __OSInitNet: DIFFERENT_FUNCTION (0.0->99.4)
- OSReturnToMenu: DIFFERENT_FUNCTION / compiler-merged private+public SDK wrap (0.0->92.5)

+++ src/system/meta/StorePackedMetadata.cpp
@@ -25,7 +25,7 @@ public:
 };
 
 void CM_CNTSDCacheClearRSO();
-int CM_CNTSDCachePushDeleteContentRSO(unsigned long, unsigned long long, unsigned short, ...);
+int CM_CNTSDCachePushDeleteContentRSO(unsigned long, unsigned long long, unsigned long, ...);
 int DebugSdBackup(unsigned long long, unsigned short);
 
 extern "C" {
